+++ internal/domain/entity/room.go
@@ -6,7 +6,7 @@ type Room struct {
 	Name           string       `json:"name"`
 	Description    string       `json:"description"`
 	Status         RoomStatus   `json:"status"`
-	RoomCategoryID uint         `json:"room_category_id"`
+	RoomCategoryID uint64       `json:"room_category_id"`
 	RoomCategory   RoomCategory `json:"room_category"`
 	Users          []User       `json:"users,omitempty"`
 }
@@ -15,14 +15,14 @@ type CreateRoom struct {
 	Name           string     `json:"name" binding:"required"`
 	Description    string     `json:"description" binding:"omitempty"`
 	Status         RoomStatus `json:"status" binding:"required,oneof=available unavailable maintenance"`
-	RoomCategoryID uint       `json:"room_category_id" binding:"required,numeric"`
+	RoomCategoryID uint64     `json:"room_category_id" binding:"required,numeric"`
 }
 
 type UpdateRoom struct {
 	Name           string     `json:"name" binding:"omitempty"`
 	Description    string     `json:"description" binding:"omitempty"`
 	Status         RoomStatus `json:"status" binding:"omitempty,oneof=available unavailable maintenance"`
-	RoomCategoryID uint       `json:"room_category_id" binding:"omitempty,numeric"`
+	RoomCategoryID uint64     `json:"room_category_id" binding:"omitempty,numeric"`
 }
 
 func (Room) TableName() string {
