+++ internal/domain/repository/otp_code_repository.go
@@ -8,5 +8,5 @@ import (
 type OtpCodeRepository interface {
 	FindCode(ctx context.Context, otpCode *entity.OtpCode) error
 	CreateOtpCode(ctx context.Context, otpCode *entity.OtpCode) error
-	UseOtpCode(ctx context.Context, otpCodeId uint) error
+	UseOtpCode(ctx context.Context, otpCodeId uint64) error
 }
