+++ internal/repository/room_repository.go
@@ -31,7 +31,7 @@ func (r *roomRepository) Create(ctx context.Context, room *entity.CreateRoom) er
 }
 
 // GetByID retrieves a room by ID
-func (r *roomRepository) GetByID(ctx context.Context, id uint) (*entity.Room, error) {
+func (r *roomRepository) GetByID(ctx context.Context, id uint64) (*entity.Room, error) {
 	var room entity.Room
 	if err := r.db.WithContext(ctx).Table(room.TableName()).Preload("RoomCategory").Preload("Users").First(&room, id).Error; err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
@@ -77,7 +77,7 @@ func (r *roomRepository) Update(ctx context.Context, room *entity.Room) error {
 }
 
 // Delete deletes a room
-func (r *roomRepository) Delete(ctx context.Context, id uint) error {
+func (r *roomRepository) Delete(ctx context.Context, id uint64) error {
 	if err := r.db.WithContext(ctx).Table(entity.Room{}.TableName()).Delete(&entity.Room{}, id).Error; err != nil {
 		return fmt.Errorf("failed to delete room: %w", err)
 	}
